Refactor datasource module to separate generators
