+++ great_expectations/datasource/__init__.py
@@ -4,4 +4,4 @@ from .sqlalchemy_source import SqlAlchemyDatasource
 from .spark_source import SparkDFDatasource
 from .dbt_source import DBTDatasource
 
-from .filesystem_path_generator import SubdirReaderGenerator, GlobReaderGenerator
+from great_expectations.datasource.generator.filesystem_path_generator import SubdirReaderGenerator, GlobReaderGenerator
